requirements.txt
pandas~=1.4.3
numpy~=1.23.0
plotly~=5.9.0
dash~=2.5.1
requests~=2.25.1
pyyaml~=6.0
geopy~=2.2.0